requirements.txt
openai
streamlit
javalang
pdfcrowd
requests
python-dotenv 